Refactor Transformer/Conformer encoder
