+++ neural_sp/models/seq2seq/encoders/conformer.py
@@ -5,16 +5,11 @@
 
 import copy
 import logging
-import random
 import torch.nn as nn
 
-from neural_sp.models.modules.conformer_convolution import ConformerConvBlock
-from neural_sp.models.modules.positionwise_feed_forward import PositionwiseFeedForward as FFN
-from neural_sp.models.modules.relative_multihead_attention import RelativeMultiheadAttentionMechanism as RelMHA
 from neural_sp.models.seq2seq.encoders.conv import ConvEncoder
+from neural_sp.models.seq2seq.encoders.conformer_block import ConformerEncoderBlock
 from neural_sp.models.seq2seq.encoders.transformer import TransformerEncoder
-
-random.seed(1)
 
 logger = logging.getLogger(__name__)
 
@@ -123,7 +118,7 @@ class ConformerEncoder(TransformerEncoder):
                                help='bottleneck dimension in the FFN layer')
             group.add_argument('--transformer_layer_norm_eps', type=float, default=1e-12,
                                help='epsilon value for layer normalization')
-            group.add_argument('--transformer_ffn_activation', type=str, default='relu',
+            group.add_argument('--transformer_ffn_activation', type=str, default='swish',
                                choices=['relu', 'gelu', 'gelu_accurate', 'glu', 'swish'],
                                help='nonlinear activation for the FFN layer')
             group.add_argument('--transformer_param_init', type=str, default='xavier_uniform',
@@ -139,7 +134,7 @@ class ConformerEncoder(TransformerEncoder):
         group.add_argument('--transformer_enc_n_heads', type=int, default=4,
                            help='number of heads in the MHA layer for Conformer encoder')
         group.add_argument('--transformer_enc_pe_type', type=str, default='relative',
-                           choices=['relative', 'relative_xl'],
+                           choices=['relative', 'relative_xl', 'none'],
                            help='type of positional encoding for Conformer encoder')
         group.add_argument('--conformer_kernel_size', type=int, default=31,
                            help='kernel size for depthwise convolution in convolution module for Conformer encoder')
@@ -176,7 +171,7 @@ class ConformerEncoder(TransformerEncoder):
         if args.transformer_enc_clamp_len > 0:
             dir_name += '_clamp' + str(args.transformer_enc_clamp_len)
         if args.dropout_enc_layer > 0:
-            dir_name += 'droplayer' + str(args.dropout_enc_layer)
+            dir_name += '_LD' + str(args.dropout_enc_layer)
         if int(args.lc_chunk_size_left.split('_')[-1]) > 0 or int(args.lc_chunk_size_current.split('_')[-1]) > 0 \
                 or int(args.lc_chunk_size_right.split('_')[-1]) > 0:
             dir_name += '_chunkL' + args.lc_chunk_size_left + 'C' + \
@@ -186,118 +181,3 @@ class ConformerEncoder(TransformerEncoder):
             dir_name += '_LA' + str(sum(list(map(int, args.transformer_enc_lookaheads.split('_')))))
 
         return dir_name
-
-
-class ConformerEncoderBlock(nn.Module):
-    """A single layer of the Conformer encoder.
-
-    Args:
-        d_model (int): input dimension of MultiheadAttentionMechanism and PositionwiseFeedForward
-        d_ff (int): hidden dimension of PositionwiseFeedForward
-        n_heads (int): number of heads for multi-head attention
-        kernel_size (int): kernel size for depthwise convolution in convolution module
-        dropout (float): dropout probabilities for linear layers
-        dropout_att (float): dropout probabilities for attention distributions
-        dropout_layer (float): LayerDrop probability
-        layer_norm_eps (float): epsilon parameter for layer normalization
-        ffn_activation (str): nonlinear function for PositionwiseFeedForward
-        param_init (str): parameter initialization method
-        pe_type (str): type of positional encoding
-        ffn_bottleneck_dim (int): bottleneck dimension for the light-weight FFN layer
-        unidirectional (bool): pad right context for unidirectional encoding
-
-    """
-
-    def __init__(self, d_model, d_ff, n_heads, kernel_size,
-                 dropout, dropout_att, dropout_layer,
-                 layer_norm_eps, ffn_activation, param_init, pe_type,
-                 ffn_bottleneck_dim, unidirectional):
-        super(ConformerEncoderBlock, self).__init__()
-
-        self.n_heads = n_heads
-        self.fc_factor = 0.5
-
-        # first half position-wise feed-forward
-        self.norm1 = nn.LayerNorm(d_model, eps=layer_norm_eps)
-        self.feed_forward_macaron = FFN(d_model, d_ff, dropout, ffn_activation, param_init,
-                                        ffn_bottleneck_dim)
-
-        # self-attention
-        self.norm2 = nn.LayerNorm(d_model, eps=layer_norm_eps)
-        self.self_attn = RelMHA(kdim=d_model,
-                                qdim=d_model,
-                                adim=d_model,
-                                odim=d_model,
-                                n_heads=n_heads,
-                                dropout=dropout_att,
-                                param_init=param_init,
-                                xl_like=pe_type == 'relative_xl')
-
-        # conv module
-        self.norm3 = nn.LayerNorm(d_model, eps=layer_norm_eps)
-        self.conv = ConformerConvBlock(d_model, kernel_size, param_init,
-                                       causal=unidirectional)
-
-        # second half position-wise feed-forward
-        self.norm4 = nn.LayerNorm(d_model, eps=layer_norm_eps)
-        self.feed_forward = FFN(d_model, d_ff, dropout, ffn_activation, param_init,
-                                ffn_bottleneck_dim)
-        self.norm5 = nn.LayerNorm(d_model, eps=layer_norm_eps)
-
-        self.dropout = nn.Dropout(dropout)
-        self.dropout_layer = dropout_layer
-
-        self.reset_visualization()
-
-    @property
-    def xx_aws(self):
-        return self._xx_aws
-
-    def reset_visualization(self):
-        self._xx_aws = None
-
-    def forward(self, xs, xx_mask=None, pos_embs=None, u_bias=None, v_bias=None):
-        """Conformer encoder layer definition.
-
-        Args:
-            xs (FloatTensor): `[B, T, d_model]`
-            xx_mask (ByteTensor): `[B, T (query), T (key)]`
-            pos_embs (LongTensor): `[L, 1, d_model]`
-            u_bias (FloatTensor): global parameter for relative positional encoding
-            v_bias (FloatTensor): global parameter for relative positional encoding
-        Returns:
-            xs (FloatTensor): `[B, T, d_model]`
-
-        """
-        self.reset_visualization()
-
-        # LayerDrop
-        if self.dropout_layer > 0 and self.training and random.random() < self.dropout_layer:
-            return xs
-
-        # first half FFN
-        residual = xs
-        xs = self.norm1(xs)
-        xs = self.feed_forward_macaron(xs)
-        xs = self.fc_factor * self.dropout(xs) + residual  # Macaron FFN
-
-        # self-attention w/ relative positional encoding
-        residual = xs
-        xs = self.norm2(xs)
-        xs, self._xx_aws = self.self_attn(xs, xs, pos_embs, xx_mask, u_bias, v_bias)
-        xs = self.dropout(xs) + residual
-
-        # conv
-        residual = xs
-        xs = self.norm3(xs)
-        xs = self.conv(xs)
-        xs = self.dropout(xs) + residual
-
-        # second half FFN
-        residual = xs
-        xs = self.norm4(xs)
-        xs = self.feed_forward(xs)
-        xs = self.fc_factor * self.dropout(xs) + residual  # Macaron FFN
-        xs = self.norm5(xs)  # this is important for performance
-
-        return xs
